refactor: move old code to legacy folder
